requirements.txt
albumentations==1.3.0
braceexpand==0.1.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
docker-pycreds==0.4.0
ExifRead-nocycle==3.0.1
filelock==3.12.0
fire==0.4.0
flatbuffers==23.5.9
fonttools==4.39.4
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.31
humanfriendly==10.0
idna==3.4
imageio==2.28.1
img2dataset==1.41.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.14.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==1.5.3
pathtools==0.1.2
Pillow==9.5.0
promise==2.3
protobuf==3.20.3
psutil==5.9.5
pyarrow==7.0.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.30.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
sentry-sdk==1.22.2
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
supervision==0.7.0
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.12.21
webdataset==0.2.48
zipp==3.15.0
